requirements.txt
azure-functions
azure-functions-durable
azure-cosmos
azure-storage-queue
pydantic
